+++ packages/flutter_scene_editor/lib/src/controller/editor_controller.dart
@@ -2095,6 +2095,8 @@ class EditorController extends ChangeNotifier {
         ..roughnessFactor = from.roughnessFactor
         ..occlusionStrength = from.occlusionStrength
         ..normalScale = from.normalScale
+        ..parallaxScale = from.parallaxScale
+        ..parallaxSteps = from.parallaxSteps
         ..doubleSided = from.doubleSided
         ..alphaMode = from.alphaMode
         ..alphaCutoff = from.alphaCutoff